requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
distro-info==1.0
idna==3.4
oauthlib==3.2.2
PyYAML==5.3.1
requests==2.28.1
requests-oauthlib==1.3.1
tweepy==4.12.1
urllib3==1.26.13
jsonschema==3.2.0
six==1.16.0
python-consul==1.1.0